requirements.txt
langchain-google-genai
langchain    
python-dotenv
google-genai
streamlit
pandas